refactor: update SchemaObject
BREAKING CHANGE: some codecs were renamed and moved
