+++ src/schema/2.0/definitions-object.ts
@@ -1,5 +1,5 @@
 import { Dictionary } from '../../utils/types';
-import { SchemaObject, SchemaObjectCodec } from './schema-object/schema-object';
+import { SchemaObject, SchemaObjectCodec } from './schema-object';
 import { dictionary } from '../../utils/io-ts';
 
 export interface DefinitionsObject extends Dictionary<SchemaObject> {}
