+++ src/schema/2.0/response-object.ts
@@ -2,7 +2,7 @@ import { optionFromNullable } from 'io-ts-types/lib/optionFromNullable';
 import { Option } from 'fp-ts/lib/Option';
 import { ExampleObject } from './example-object';
 import { HeadersObject } from './headers-object';
-import { SchemaObject, SchemaObjectCodec } from './schema-object/schema-object';
+import { SchemaObject, SchemaObjectCodec } from './schema-object';
 import { string, type } from 'io-ts';
 
 export interface ResponseObject {
