--- src/schema/2.0/schema-object/boolean-property-schema-object.ts
@@ -1,12 +1,0 @@
-import { literal, type } from 'io-ts';
-
-export interface BooleanPropertySchemaObject {
-	readonly type: 'boolean';
-}
-
-export const BooleanPropertySchemaObject = type(
-	{
-		type: literal('boolean'),
-	},
-	'BooleanPropertySchemaObject',
-);
